homepage: separate component for how it works section
