--- frontend/containers/home/footer-banner/types.ts
@@ -1,6 +1,0 @@
-import { Project } from 'types/project';
-
-export type ImageGalleryProps = {
-  /** images array */
-  images: Project['project_images'];
-};
