+++ frontend/containers/home/how-it-works/index.ts
